pax.c, awk_parse.c: clear the lint (warn) -Wsign-compare backlog
Three genuinely distinct sites, each safe by invariant rather than a
real bug:

- ustar_put_oct()/cpio_put_field() (pax.c): both loop over `width - 1`/
  `width`, an int, into a size_t index. Every call site passes a
  literal 8 or 12 (ustar) or is already contracted to __arith_range(6,
  11) and proven at every call site by the arithub lint stage (cpio).
  width itself has to stay `int` -- snprintf's "%0*lo" requires an int
  argument for the field-width `*` -- so the fix is a cast at the loop
  bound, matching the (size_t) casts the surrounding
  __ownership_*_span() calls already use for the same reason.

- assign_ops[] (awk_parse.c): declared `const int`, holding
  awk_toktype enumerators plus a trailing 0 sentinel, and compared
  directly against p->tok.type (enum awk_toktype). All enumerators are
  non-negative, so this is not a bug, but the array's own declared
  type was wrong -- fixed at the source by declaring it `const enum
  awk_toktype` (0 renamed to T_EOF, which it already denoted, to keep
  the sentinel self-documenting under the new type).

Verified against the real lint (warn) invocation (WARN_FLAGS drawn
from tools/lint.sh's stage_warn) across gcc and clang on all three
arches: 18 raw -Wsign-compare hits (12 pax.c, 6 awk_parse.c -- 3
unique sites x 6 (arch, compiler) passes each) drop to 0, with no
newly introduced warnings. No other file-wide diagnostic appears for
either file post-fix. test/util-archive.c and test/util-awk.c pass
unchanged.

+++ src/util/pax.c
@@ -219,7 +219,10 @@ static void ustar_put_oct(unsigned char *field, int width, unsigned long value)
 	__ownership_readable_span(tmp, (size_t)(width - 1));
 	{
 		size_t i;
-		for (i = 0; i < width - 1; i++) field[i] = tmp[i];
+		/* Every caller below passes a literal 8 or 12; the cast mirrors
+		 * the two __ownership_*_span() calls just above, which already
+		 * treat width - 1 as a size_t for the same reason. */
+		for (i = 0; i < (size_t)(width - 1); i++) field[i] = tmp[i];
 	}
 	field[width - 1] = 0;
 }
@@ -372,7 +375,9 @@ static int cpio_put_field(char *field, int width __arith_range(6, 11),
 	}
 	snprintf(tmp, sizeof tmp, "%0*lo", width, value);
 	__ownership_writable_span(field, (size_t)width);
-	for (size_t i = 0; i < width; i++) field[i] = tmp[i];
+	/* width's __arith_range(6, 11) above is proven at every call site by
+	 * the arithub lint stage, so this is never negative. */
+	for (size_t i = 0; i < (size_t)width; i++) field[i] = tmp[i];
 	return 0;
 }
